Rework how online and automatic learning occurs
Automatic learning (from users clicking on the 3 choices) is too noisy
to be mixed up with manual training from the website. Instead,
separate it into different datasets.

Furthermore, there is no point in keeping a test dataset from online
learning. Instead, it's worthwhile to just log all utterances that
go through the server, and in the future use data programming on
them. This is disabled by default but can be enabled by default
for the production instance, and can also be controlled on a sentence
by sentence basis.